Organize H5 integration tests under build outputs
