+++ .github/workflows/pr.yaml
@@ -185,7 +185,7 @@ jobs:
           tests/integration/test_tiny_stage_5_artifacts.py
           tests/integration/test_tiny_pipeline_e2e.py
           tests/integration/test_tiny_pipeline_h5_e2e.py
-          tests/integration/local_h5/
+          tests/integration/build_outputs/
           tests/integration/test_modal_pipeline_seams.py
           tests/integration/test_tiny_h5_pipeline.py
           tests/integration/test_modal_pipeline_e2e.py
