+++ tests/integration/build_outputs/h5_traceability/test_traceability_contract.py
@@ -4,7 +4,11 @@ from policyengine_us_data.build_outputs.fingerprinting import (
 )
 import pytest
 
-from tests.integration.local_h5.fixtures import SEED, VERSION, seed_local_h5_artifacts
+from tests.integration.build_outputs.fixtures import (
+    SEED,
+    VERSION,
+    seed_local_h5_artifacts,
+)
 
 pytestmark = pytest.mark.integration
